# SIMD version of BLEND_RGBA_MULT for 32bit surfaces