--- a/src_c/alphablit.c
+++ b/src_c/alphablit.c
@@ -289,36 +289,36 @@ SoftBlitPyGame(SDL_Surface *src, SDL_Rect *srcrect, SDL_Surface *dst,
                 case PYGAME_BLEND_RGBA_MULT: {
 #if defined(HAVE_IMMINTRIN_H) && !defined(SDL_DISABLE_IMMINTRIN_H)
                     if (src->format->BytesPerPixel == 4 &&
-                            dst->format->BytesPerPixel == 4 &&
-                            src->format->Rmask == dst->format->Rmask &&
-                            src->format->Gmask == dst->format->Gmask &&
-                            src->format->Bmask == dst->format->Bmask &&
-                            info.src_blend != SDL_BLENDMODE_NONE &&
-                            SDL_HasAVX2()) {
+                        dst->format->BytesPerPixel == 4 &&
+                        src->format->Rmask == dst->format->Rmask &&
+                        src->format->Gmask == dst->format->Gmask &&
+                        src->format->Bmask == dst->format->Bmask &&
+                        info.src_blend != SDL_BLENDMODE_NONE &&
+                        SDL_HasAVX2()) {
                         blit_blend_rgba_mul_simd_avx2(&info);
                         break;
                     }
 #endif /* defined(HAVE_IMMINTRIN_H) && !defined(SDL_DISABLE_IMMINTRIN_H) */
 #if defined(__SSE2__)
                     if (src->format->BytesPerPixel == 4 &&
-                            dst->format->BytesPerPixel == 4 &&
-                            src->format->Rmask == dst->format->Rmask &&
-                            src->format->Gmask == dst->format->Gmask &&
-                            src->format->Bmask == dst->format->Bmask &&
-                            info.src_blend != SDL_BLENDMODE_NONE &&
-                            SDL_HasSSE2()) {
+                        dst->format->BytesPerPixel == 4 &&
+                        src->format->Rmask == dst->format->Rmask &&
+                        src->format->Gmask == dst->format->Gmask &&
+                        src->format->Bmask == dst->format->Bmask &&
+                        info.src_blend != SDL_BLENDMODE_NONE &&
+                        SDL_HasSSE2()) {
                         blit_blend_rgba_mul_simd(&info);
                         break;
                     }
 #endif /* __SSE2__*/
 #if PG_ENABLE_ARM_NEON
                     if (src->format->BytesPerPixel == 4 &&
-                            dst->format->BytesPerPixel == 4 &&
-                            src->format->Rmask == dst->format->Rmask &&
-                            src->format->Gmask == dst->format->Gmask &&
-                            src->format->Bmask == dst->format->Bmask &&
-                            info.src_blend != SDL_BLENDMODE_NONE &&
-                            SDL_HasNEON()) {
+                        dst->format->BytesPerPixel == 4 &&
+                        src->format->Rmask == dst->format->Rmask &&
+                        src->format->Gmask == dst->format->Gmask &&
+                        src->format->Bmask == dst->format->Bmask &&
+                        info.src_blend != SDL_BLENDMODE_NONE &&
+                        SDL_HasNEON()) {
                         blit_blend_rgba_mul_simd(&info);
                         break;
                     }
@@ -2593,7 +2593,7 @@ alphablit_alpha_sse2_argb_no_surf_alpha(SDL_BlitInfo *info)
             srcp64 += srcskip;
             dstp64 += dstskip;
         }
-#else  /* 32 bit 2 pixel path */
+#else /* 32 bit 2 pixel path */
 
         /* two pixels at a time - 32 bit version - only works when blit width
            is an even number */
@@ -2918,7 +2918,7 @@ alphablit_alpha_sse2_argb_no_surf_alpha_opaque_dst(SDL_BlitInfo *info)
             srcp64 += srcskip;
             dstp64 += dstskip;
         }
-#else  /* 32 bit */
+#else /* 32 bit */
 
         /* two pixels at a time - 32 bit version - only works when blit width
            is an even number */