--- a/src_c/alphablit.c
+++ b/src_c/alphablit.c
@@ -2918,7 +2918,7 @@ alphablit_alpha_sse2_argb_no_surf_alpha_opaque_dst(SDL_BlitInfo *info)
             srcp64 += srcskip;
             dstp64 += dstskip;
         }
-#else /* 32 bit */
+#else
 
         /* two pixels at a time - 32 bit version - only works when blit width
            is an even number */
@@ -2987,7 +2987,7 @@ alphablit_alpha_sse2_argb_no_surf_alpha_opaque_dst(SDL_BlitInfo *info)
             srcp64 += srcskip;
             dstp64 += dstskip;
         }
-#endif /* 32 bit */
+#endif
     }
     else {
         /* one pixel at a time */