--- a/src_c/simd_blitters_sse2.c
+++ b/src_c/simd_blitters_sse2.c
@@ -113,7 +113,7 @@ blit_blend_rgba_mul_sse2(SDL_BlitInfo *info)
     }
 }
 #else
-static void
+void
 blit_blend_rgba_mul_sse2(SDL_BlitInfo *info)
 {
     int n;