This PR explorers adding a SIMD version of the `BLEND_RGBA_MULT` blend mode. The goal is to speed up this blend mode and hopefully provide a template for speeding up the other special blend modes such as `BLEND_RGBA_ADD` and `BLEND_RGBA_MAX` with SIMD techniques.

**Background**

The special blend modes are used by supplying one of them to the `special_flag` parameter of blit, you can see a list of them in the documentation for `blit()` here:
https://www.pygame.org/docs/ref/surface.html#pygame.Surface.blit

They are useful for various effects, but up until now have been implemented entirely using standard C maths, generally doing the operations on a colour channel at a time. This is in contrast to our normal blitters and the `BLEND_PREMULTIPLIED_ALPHA` blitter which all use some for of SIMD acceleration either via SDL, or our own handwritten SIMD instructions.

Generally the existing SIMD blitters in pygame are written in SSE2 compliant intrinsic functions as these have wide support and have existed in most computers since around the year 2000. There is potential for speeding things up further with newer intrinsic standards (though those will have less widespread support), but I think it is best to at least get us into the year 2000 before we start reaching for 2022.

Arm processers have their own, similar intrinsic functions for speeding up these kinds of math operations and at the minute we use a library called `sse2neon` from here:
https://github.com/DLTcollab/sse2neon

To automatically convert our SSE2 intrinsic functions to arm neon versions. I suspect the porting is not as fast as it could be with native arm instructions written but I personally have no experience with arm intrinsic functions or a good setup to test them. In my experience programming with intrinsic functions is already fairly confusing, which is part of why I'm writing a lot in this PR to try and explain what I'm doing.

**SSE2 SIMD intrinsic functions**

The clearest guide to these is from intel here:
https://www.intel.com/content/www/us/en/docs/intrinsics-guide/index.html

You can search for individual functions with the search box and tick the standards at the side to narrow down what intrinsic standard you wish to support up to. Generally newer standards are inclusive of older ones ( I think always?) so a processor supporting SSE2 also supports SSE and MMX.

The MMX standard supported 64 bit memory registers, while the SSE standard supports 128bit registers. If we were to create AVX SIMD versions we would have 256bit memory registers and the newest AV-512 standard supports, you guessed it 512 bit memory registers.

The point of these large memory registers and SIMD in general, is to run simultaneous functions on multiple parts of the large registers at the same time. For example you might have a function that does an add operation 8 times on each of the 8, 16 bit chunks of a 128bit register in one call. This is perfect for pixel operations as each colour channel is only an 8 bit number between 0 to 255 (or from 0x00 to 0xFF to put it in hexadecimal). Some operations are also saturating so if you add two 8 bit numbers together with an 8 bit adding function it will clamp the result to a max of 255 for you.

Running operations on data in the registers is very fast but loading data from normal memory into the SIMD registers can be quite slow. This is why you need to carefully profile to see if it is actually worth spending a lot of time loading multiple pixels worth of data into a SIMD register before running an operation on it.

Further complicating matters, some scalar math code gets automatically vectorised by compilers so you may be competing against a compiler that has already done some obvious optimisations in basic maths.

**This PR specifically** 

As I said at the beginning I want to create a simple template for SIMD versions of the special blend modes.

The last time I worked on SIMD it was suggested it might be better to move the SIMD functions with their extra imports and compiler demands into a separate file, so I thought I would open with an attempt to do this. This is why the new `blit_blend_rgba_mul_simd` is in its own header file - I'm not sure it is possible to stick it into a .c file and still have it be static and called from `alphablit.c` but having their own header is something at least. If this approach seems worthwhile more of the existing SIMD blitters could move into here instead of in `alphablit.c` . I also moved the blit_info structure into it's own header file which may not be necessary without a new .c file for the SIMD blitters but it tidied up the top of `alphablit.c` a bit so I left it.

Some extra if checks were then added to push SSE2 supporting processors handing a 32 bit to 32 bit surface blit over to the SIMD version of the blending function.

Then we get to the meat of the blending function. To make the blend work in a reasonable SIMD way I tweaked the formula for the blend. Both the existing formula and my new version are approximations of this operation:

```
dst = (dst * src) / 255
```
which is a standard multiply blend. However divide is too expensive an operation to do four times for every colour channel of every pixel so the existing formula for 32 bit surfaces is actually this:

```
tmp = (dst && src) ? (dst * src) >> 8 : 0
dst = (tmp <= 255 ? tmp : 255);
```

which uses a bitshift to approximate divide and branches to try and clamp the result between 0 and 255. I'm fairly sure that neither of these branches is necessary and the bit shift approximation could be better because it is prone to an off by one issue. If you do:

```
(255 * 255) >> 8
```

You will see that the result is 254 not 255 as you would expect from `dst = (dst * src) / 255` likewise:

```
((1 * 255)) >> 8
```
gives a result of 0 rather than 1.

Anyway, its just an approximation so it's not too bothersome but it could be better and the clamps between 0, and 255 are unnecessary as the max & min values you can inject into the bitshift calculation evaluate to 0 and 254. 

So, the first step of my SIMD journey is to fix this calculation up to be more accurate and faster. I propose we use this instead as the underlying operation:

```
dst = ((dst * src) + 255) >> 8
```

Adding 255 to the result of the multiplication fixes the off by one errors of the bitshift:

```
>>>((255 * 255) + 255) >> 8
255
>>>((1 * 255) + 255) >> 8
1
>>>((0 * 255) + 255) >> 8
0
>>>((128 * 200) + 255) >> 8 
100
```
And there is no need for any ternary operator branches because the result is already bound between 0 and 255. Withotu any branches it is also easy to vectorise with SIMD.

So far I've created a one-pixel-at-a-time version that:

 - first copies a single 32 bit pixels worth of data from the normal memory of the src and dst surfaces into two SIMD registers.
 - then spreads each 8 bit colour channel out over 16 bits so there is room for a multiplication (multiplying two 8 bit numbers together requires 16 bits of memory room for the result)  
 - then does the 16bit multiplication of the two SIMD registers, effectively multiplying each colour channel simultaneously.
 - then adds 255 onto the result of each channels multiply using another SIMD register filled with spread out 0xFF (255).
 - then does the bitshift on all the channels completing the blend.
 - finally we repack the "16 bit" channels back down to 8bit (they only have 8bit data at this point after the bit shift, it is just spread out)
 - then we take the new 32bit pixel out of the SIMD register back into normal memory. 

And that's it.

The next step is to see if I can make a version that does two pixels at a time, and if that version is actually any faster or if the time spend loading pixels in actually makes it slower than one pixel at a time. Doing 2 pixels at a time also requires an even width blit to function in the current blitting paradigm which complicates matters further.

**Performance results so far**

The only reason to do this is if it speeds up the blitter so performance data is important. I made a test program which does blits at various sizes with picked blend modes over a number of iterations and times how long it takes. Here is the program:

```
import time
import math
import pygame


def test(target_surf, test_surf, methods, tries, blit_pos):
    for method in methods:
        start = time.time()
        name = method(target_surf, test_surf, tries, blit_pos)
        results = (time.time() - start) * 1000
        name_tab_length = math.ceil(len('tested {}:'.format(name)) / 4)
        tabs = '\t' * (8 - name_tab_length)
        string_to_format = 'tested {}:' + tabs + '{}ms'
        print(string_to_format.format(name, round(results, 3)))


def blend_rgba_mult(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_RGBA_MULT)
    return 'Blit BLEND_RGBA_MULT'


def run_rgba32_blit_test(target_surf, iterations, width, height, blend_mode_funcs):
    test_name = 'Blend RGBA 32 Test'
    test_surf = pygame.Surface((width, height), flags=pygame.SRCALPHA, depth=32)
    test_surf.fill((100, 100, 100, 100))
    num_tries = iterations
    start_pos = ((screen.get_width() - width), screen.get_height() - height)
    odd_start_pos = ((screen.get_width() - width) + 1, screen.get_height() - height)
    print(f'\nTest type - {test_name}: {num_tries} blits of image size {width}, {height}')
    test(target_surf, test_surf, blend_mode_funcs, num_tries, start_pos)


pygame.init()
screen = pygame.display.set_mode((1280, 1024))

surf_to_blit_onto = pygame.Surface((1280, 1024))
surf_to_blit_onto.fill((50, 50, 50))

test_iterations = 1000
blend_modes_to_test = [blend_rgba_mult]  # (blend_add, blend_rgb_add, blend_mult, blend_rgb_mult)

run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 32, 32, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 64, 64, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 128, 128, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 256, 256, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 512, 512, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 1024, 1024, blend_modes_to_test)

``` 

Which produces the following results on 2.1.2 on my windows PC:

```
Test type - Blend RGBA 32 Test: 1000 blits of image size 32, 32
tested Blit BLEND_RGBA_MULT:	13.012ms

Test type - Blend RGBA 32 Test: 1000 blits of image size 64, 64
tested Blit BLEND_RGBA_MULT:	48.043ms 

Test type - Blend RGBA 32 Test: 1000 blits of image size 128, 128
tested Blit BLEND_RGBA_MULT:	191.175ms

Test type - Blend RGBA 32 Test: 1000 blits of image size 256, 256
tested Blit BLEND_RGBA_MULT:	761.176ms

Test type - Blend RGBA 32 Test: 1000 blits of image size 512, 512
tested Blit BLEND_RGBA_MULT:	3036.053ms

Test type - Blend RGBA 32 Test: 1000 blits of image size 1024, 1024
tested Blit BLEND_RGBA_MULT:	12022.943ms
```

Then after this PR in it's current state we get:

```
Test type - Blend RGBA 32 Test: 1000 blits of image size 32, 32
tested Blit BLEND_RGBA_MULT:	1.002ms = 13x faster

Test type - Blend RGBA 32 Test: 1000 blits of image size 64, 64
tested Blit BLEND_RGBA_MULT:	4.003ms = 12x faster

Test type - Blend RGBA 32 Test: 1000 blits of image size 128, 128
tested Blit BLEND_RGBA_MULT:	12.012ms = 15x faster

Test type - Blend RGBA 32 Test: 1000 blits of image size 256, 256
tested Blit BLEND_RGBA_MULT:	48.043ms = 15x faster

Test type - Blend RGBA 32 Test: 1000 blits of image size 512, 512
tested Blit BLEND_RGBA_MULT:	190.173ms = 15x faster

Test type - Blend RGBA 32 Test: 1000 blits of image size 1024, 1024
tested Blit BLEND_RGBA_MULT:	750.235ms = 16x faster
```

So between 12 and 16 times faster at all the different blit sizes so far. 

I'll update this draft as I try other things with this blend mode until I think I've got a satisfactory template for 32bit blits for the other RGBA_ special blend modes (MAX,MIN,SUB & ADD). To keep it (relatively) simple I'll add changes for those blend modes in other PRs later as I think this topic is complicated enough already.

Let me know if you have any ideas, feedback or thoughts. I'm sure there is lots of stuff I missed as this is a complex topic and I forget half of it each time I go away from it and come back after a break.