@MyreMylar 2022-02-07
After some further experimentation I have:

 - **Determined that 'stride switching' works much better than expected.** This is where blits with widths that don't fit neatly into an '8 or 2 pixels at a time' blits do the excess pixels with a 'one-pixel-at-a-time' algorithm before switching to the 8 or 2 pixel version afterwards. Performance was easily comparable with the width switch version I had before in the optimum case and much better for the non-optimum cases. As a result all version of the blitter use this stride switch approach and average blit performance (assuming widths are random) is improved. 
 - **Upgraded the SSE2 version to do 2 pixels at a time when it can.** This unsurprisingly boosts it's performance.
 - **Implemented an AVX2 version of the blitter.** This is about as far into the SIMD tech as we can go at the minute and should get good usage (86.31% in the January Steam hardware survey - compared to 100% for SSE2). AVX-512 is not a usable technology at present ( there are a ton of sub-versions, intel has just dropped it, AMD hasn't added it yet and it only has about 5% uptake)
 - **Created a new script to test 'real-world' usage of the blitter.** I suspected the stats from my speed test were being inflated by the test being a 'perfect' setup for SIMD which was unlikely to occur in a real program. This proved to be true, the new blitters are faster - but not absurdly faster in a more real world scenario.
 - **Expanded the speed test to check very small (sub 4x4) blits in case they were slower**. They are not slower - but they also aren't any faster, unsurprisingly blitting many pixels at a time works better when there are many pixels to blit.

Here is my updated pure speed test script:
```
import time
import math
import pygame


def test(target_surf, test_surf, methods, tries, blit_pos, extra_desc):
    for method in methods:
        start = time.time()
        name = method(target_surf, test_surf, tries, blit_pos)
        results = (time.time() - start) * 1000
        desc_string_start = 'tested {}' + extra_desc + ':'
        name_tab_length = math.ceil((len(desc_string_start.format(name)) + 1) / 4)
        tabs = '\t' * (19 - name_tab_length)
        string_to_format = desc_string_start + tabs + '{}ms'
        print(string_to_format.format(name, round(results, 3)))


def blend_add(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_ADD)
    return 'Blit BLEND_ADD'


def blend_rgb_add(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_RGB_ADD)
    return 'Blit BLEND_RGB_ADD'


def blend_mult(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_MULT)
    return 'Blit BLEND_MULT'


def blend_rgb_mult(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_RGB_MULT)
    return 'Blit BLEND_RGB_MULT'


def blend_rgba_mult(target_surf, blit_surf, tries, blit_pos):
    for t in range(tries):
        target_surf.blit(blit_surf, blit_pos, special_flags=pygame.BLEND_RGBA_MULT)
    return 'Blit BLEND_RGBA_MULT'


def run_rgba32_blit_test(target_surf, iterations, width, height, blend_mode_funcs):
    test_name = 'Blend RGBA 32 Test'
    test_surf = pygame.Surface((width, height), flags=pygame.SRCALPHA, depth=32)
    test_surf.fill((100, 100, 100, 100))
    num_tries = iterations
    start_pos = ((screen.get_width() - width), screen.get_height() - height)
    odd_start_pos = ((screen.get_width() - width) + 1, screen.get_height() - height)
    # print(f'\nTest type - {test_name}: {num_tries} blits of image size {width}, {height}')
    test(target_surf, test_surf, blend_mode_funcs,
         num_tries, start_pos, f' - {num_tries} tries, even dimensions {width}, {height} ')
    test(target_surf, test_surf, blend_mode_funcs,
         num_tries, odd_start_pos, f' - {num_tries} tries, odd dimensions {width-1}, {height} ')


pygame.init()
screen = pygame.display.set_mode((1280, 1024))

surf_to_blit_onto = pygame.Surface((1280, 1024))
surf_to_blit_onto.fill((50, 50, 50))

test_iterations = 1000
blend_modes_to_test = [blend_rgba_mult]  # (blend_add, blend_rgb_add, blend_mult, blend_rgb_mult)

# standard size ranges
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 32, 32, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 64, 64, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 128, 128, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 256, 256, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 512, 512, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 1024, 1024, blend_modes_to_test)

# very small size ranges
test_iterations = 50000
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 16, 16, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 8, 8, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 4, 4, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 2, 2, blend_modes_to_test)
run_rgba32_blit_test(surf_to_blit_onto, test_iterations, 1, 1, blend_modes_to_test)

```

And here is my 'real-world' test script that switches on a full-screen 'post-process' vignette and noise when you press the return key:

```
from collections import deque
import random
import pygame


def generate_random_white_noise(size):
    noise_surf = pygame.Surface(size, flags=pygame.SRCALPHA)
    noise_surf.fill((255, 255, 255, 255))

    for x in range(0, size[0]):
        for y in range(0, size[1]):
            noise_val = random.gauss(0.975, 0.008)
            rand_lum = min(255, max(0, int(noise_val * 255.0)))
            noise_surf.set_at((x, y), pygame.Color(rand_lum, rand_lum, rand_lum, 255))
    return noise_surf


pygame.init()
window_size = (1920, 1080)
window_surf = pygame.display.set_mode(window_size)
background_surf = pygame.Surface(window_size)
background_surf.fill((0, 0, 0))
game_level_image = pygame.image.load('test_images/game_level_test.jpg').convert()
game_level_image = pygame.transform.smoothscale(game_level_image, window_size)

vignette_small = pygame.Surface((240, 135))
vignette_small.fill((0, 0, 0))
pygame.draw.rect(vignette_small, pygame.Color((32, 32, 32)),
                 pygame.Rect((-20, -20), (280, 175)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((64, 64, 64)),
                 pygame.Rect((-15, -15), (270, 165)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((96, 96, 96)),
                 pygame.Rect((-10, -10), (260, 155)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((128, 128, 128)),
                 pygame.Rect((-5, -5), (250, 145)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((160, 160, 160)),
                 pygame.Rect((0, 0), (240, 135)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((192, 192, 192)),
                 pygame.Rect((5, 5), (230, 125)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((224, 224, 224)),
                 pygame.Rect((10, 10), (220, 115)), border_radius=128)
pygame.draw.rect(vignette_small, pygame.Color((255, 255, 255)),
                 pygame.Rect((15, 15), (210, 105)), border_radius=128)

vignette_small = pygame.transform.smoothscale(vignette_small, (60, 32))
vignette = pygame.transform.smoothscale(vignette_small, window_size)

noise_anim = [pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size),
              pygame.transform.smoothscale(generate_random_white_noise((480, 270)), window_size)]
noise_index = 0
noise = noise_anim[noise_index]
noise_acc = 0.0
noise_frametime = 0.05

font = pygame.font.Font(None, 64)

time_delta_stack = deque([])
display_vignette = False
clock = pygame.time.Clock()
running = True
fps_text = "FPS: "
average_time_delta = 0.0
while running:
    time_delta = clock.tick()/1000.0

    for event in pygame.event.get():
        if event.type == pygame.QUIT:
            running = False
        if event.type == pygame.KEYDOWN:
            if event.key == pygame.K_ESCAPE:
                running = False
            if event.key == pygame.K_RETURN:
                if display_vignette:
                    display_vignette = False
                else:
                    display_vignette = True

    noise_acc += time_delta
    if noise_acc >= noise_frametime:
        noise_acc = 0.0
        noise_index += 1
        if noise_index == len(noise_anim):
            noise_index = 0
        noise = noise_anim[noise_index]

    window_surf.blit(background_surf, (0, 0))
    window_surf.blit(game_level_image, (0, 0))
    if display_vignette:
        window_surf.blit(noise, (0, 0), special_flags=pygame.BLEND_RGBA_MULT)
        window_surf.blit(vignette, (0, 0), special_flags=pygame.BLEND_RGBA_MULT)

    time_delta_stack.append(time_delta)
    if len(time_delta_stack) > 100:
        time_delta_stack.popleft()

    if len(time_delta_stack) == 100:
        average_time_delta = sum(time_delta_stack) / 100.0
        fps_text = f'FPS: {min(999.0, 1.0 / max(average_time_delta, 0.0000001)):.2f}'
        fps_surf = font.render(fps_text, True, pygame.Color('#FFFFFF'))
        window_surf.blit(fps_surf, (1650, 80))

    pygame.display.update()

print(fps_text)
print("time delta: ", average_time_delta)

```

I used this test image:
![game_level_test](https://user-images.githubusercontent.com/13382426/152860096-a74a731b-3ec7-4f1f-b908-58bd5be8f1b1.jpg)

Which looks like this with the post-process running:
![image](https://user-images.githubusercontent.com/13382426/152860293-7a72a9b1-396f-4a53-b5bb-afb693f95c4d.png)

Here are the results from the speed tests and real-world tests on my windows PC with an AVX2 capable processor (AMD Ryzen 5 3600X):

**Current (pygame 2.1.2):**

```
OLD
----
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 32, 32 :		13.012ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 31, 32 :		12.011ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 64, 64 :		50.046ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 63, 64 :		48.044ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 128, 128 :		190.172ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 127, 128 :		187.171ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 256, 256 :		747.405ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 255, 256 :		740.675ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 512, 512 :		2986.622ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 511, 512 :		2984.607ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 1024, 1024 :		11956.732ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 1023, 1024 :		11858.296ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 16, 16 :		178.633ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 15, 16 :		167.445ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 8, 8 :		63.527ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 7, 8 :		59.053ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 4, 4 :		36.033ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 3, 4 :		32.03ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 2, 2 :		27.025ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 1, 2 :		26.023ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 1, 1 :		25.023ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 0, 1 :		21.019ms

Real world test:
----------------
FPS: 28.77
time delta:  0.03476000000000002
```

**This branch with SSE 2 only (AVX2 disabled):**

```
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 32, 32 :		1.001ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 31, 32 :		1.001ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 64, 64 :		2.002ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 63, 64 :		2.09ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 128, 128 :		6.917ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 127, 128 :		7.007ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 256, 256 :		26.023ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 255, 256 :		28.026ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 512, 512 :		104.094ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 511, 512 :		107.097ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 1024, 1024 :		408.841ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 1023, 1024 :		425.025ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 16, 16 :		30.982ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 15, 16 :		30.992ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 8, 8 :		26.005ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 7, 8 :		28.026ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 4, 4 :		24.977ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 3, 4 :		25.996ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 2, 2 :		25.005ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 1, 2 :		25.004ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 1, 1 :		23.995ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 0, 1 :		21.001ms

Real world test:
----------------
FPS: 147.06
time delta:  0.006800000000000005
```

**This branch with AVX2:**

```
AVX2
----
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 32, 32 :		1.0ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 31, 32 :		0.0ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 64, 64 :		2.002ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 63, 64 :		1.001ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 128, 128 :		3.003ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 127, 128 :		4.004ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 256, 256 :		13.012ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 255, 256 :		14.013ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 512, 512 :		44.04ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 511, 512 :		48.043ms
tested Blit BLEND_RGBA_MULT - 1000 tries, even dimensions 1024, 1024 :		172.158ms
tested Blit BLEND_RGBA_MULT - 1000 tries, odd dimensions 1023, 1024 :		176.678ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 16, 16 :		28.999ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 15, 16 :		32.0ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 8, 8 :		26.999ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 7, 8 :		28.007ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 4, 4 :		25.994ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 3, 4 :		25.006ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 2, 2 :		25.988ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 1, 2 :		25.014ms
tested Blit BLEND_RGBA_MULT - 50000 tries, even dimensions 1, 1 :		25.915ms
tested Blit BLEND_RGBA_MULT - 50000 tries, odd dimensions 0, 1 :		21.082ms

Real world test:
----------------
FPS: 149.93
time delta:  0.006670000000000005
```

As you can see, while AVX2 looks great in the speed benchmark, it is only marginally better than SSE2 in the real world test. Both SIMD blitters, in both tests, seem to be a bit improvement over the current non-SIMD version. 

Assuming most blits are not supper small or super large and that real world blitting is not going to be as good as the speed benchmark - perhaps it comes out to about 5 times faster?

I'd like to see this tested on ARM (needs SIMD enabled) and I guess linux and mac (GCC rather than MSVC compiler) to see if it get similar results. 

I also clearly need to figure out how to install clang to get his passing the CI.